fix(nix/dependency-analyzer): fix check for readFile with context
This is a nixVersion check that isn't bogus which is more involved to do
correctly in the current situation. There's no cheap way to check for
this feature, so using the version is still the best option.

- We know that Lix has it and Lix doesn't intend to change its nixVersion
  string, so that's easy.
- We know C++ Nix >= 2.6 has it, but we need to make sure that we don't
  get fooled by the //3p/cppnix user agent string. We currently also
  exclude Tvix. In both cases, we can adjust this condition once we know
  they support this feature.

This should restore dependency information to the pipelines which have
been missing recently.

Change-Id: I09b8d3ede12957bc6232d649d2aa1af8bb228c5c
Reviewed-on: https://cl.tvl.fyi/c/depot/+/13836
Reviewed-by: tazjin <tazjin@tvl.su>
Tested-by: BuildkiteCI
Autosubmit: sterni <sternenseemann@systemli.org>